requirements.txt

numpy<=1.26
torch>=2.2
ultralytics
opencv-python
pillow